优化编译脚本

+++ build.sh
@@ -7,26 +7,24 @@ PROJECT="Jrohy/trojan"
 #获取当前的这个脚本所在绝对路径
 SHELL_PATH=$(cd `dirname $0`; pwd)
 
-RELEASE_ID=`curl -H 'Cache-Control: no-cache' -s https://api.github.com/repos/$PROJECT/releases/latest|grep id|awk 'NR==1{print $2}'|sed 's/,//'`
-
 function uploadfile() {
-  FILE=$1
-  CTYPE=$(file -b --mime-type $FILE)
+	FILE=$1
+	CTYPE=$(file -b --mime-type $FILE)
 
-  curl -H "Authorization: token ${GITHUB_TOKEN}" -H "Content-Type: ${CTYPE}" --data-binary @$FILE "https://uploads.github.com/repos/$PROJECT/releases/${RELEASE_ID}/assets?name=$(basename $FILE)"
+	curl -H "Authorization: token ${GITHUB_TOKEN}" -H "Content-Type: ${CTYPE}" --data-binary @$FILE "https://uploads.github.com/repos/$PROJECT/releases/${RELEASE_ID}/assets?name=$(basename $FILE)"
 
-  echo ""
+	echo ""
 }
 
 function upload() {
-  FILE=$1
-  DGST=$1.dgst
-  openssl dgst -md5 $FILE | sed 's/([^)]*)//g' >> $DGST
-  openssl dgst -sha1 $FILE | sed 's/([^)]*)//g' >> $DGST
-  openssl dgst -sha256 $FILE | sed 's/([^)]*)//g' >> $DGST
-  openssl dgst -sha512 $FILE | sed 's/([^)]*)//g' >> $DGST
-  uploadfile $FILE
-  uploadfile $DGST
+	FILE=$1
+	DGST=$1.dgst
+	openssl dgst -md5 $FILE | sed 's/([^)]*)//g' >> $DGST
+	openssl dgst -sha1 $FILE | sed 's/([^)]*)//g' >> $DGST
+	openssl dgst -sha256 $FILE | sed 's/([^)]*)//g' >> $DGST
+	openssl dgst -sha512 $FILE | sed 's/([^)]*)//g' >> $DGST
+	uploadfile $FILE
+	uploadfile $DGST
 }
 
 cd $SHELL_PATH
@@ -40,17 +38,25 @@ LDFLAGS="-w -s -X 'trojan/trojan.MVersion=$VERSION' -X 'trojan/trojan.BuildDate=
 GOOS=linux GOARCH=amd64 go build -ldflags "$LDFLAGS" -o "result/trojan-linux-amd64" .
 GOOS=linux GOARCH=arm64 go build -ldflags "$LDFLAGS" -o "result/trojan-linux-arm64" .
 
-cd result
+if [[ $# == 0 ]];then
+	cd result
 
-UPLOAD_ITEM=($(ls -l|awk '{print $9}'|xargs -r))
+	UPLOAD_ITEM=($(ls -l|awk '{print $9}'|xargs -r))
 
-for ITEM in ${UPLOAD_ITEM[@]}
-do
-   upload $ITEM
-done
+	curl -X POST -H "Authorization: token ${GITHUB_TOKEN}" -H "Accept: application/vnd.github.v3+json" https://api.github.com/repos/$PROJECT/releases -d '{"tag_name":"'$VERSION'", "name":"'$VERSION'"}'
 
-echo "upload completed!"
+	sleep 2
 
-cd $SHELL_PATH
+	RELEASE_ID=`curl -H 'Cache-Control: no-cache' -s https://api.github.com/repos/$PROJECT/releases/latest|grep id|awk 'NR==1{print $2}'|sed 's/,//'`
 
-rm -rf result
+	for ITEM in ${UPLOAD_ITEM[@]}
+	do
+		upload $ITEM
+	done
+
+	echo "upload completed!"
+
+	cd $SHELL_PATH
+
+	rm -rf result
+fi
